XW-1114 | move custom errors to a separate file
